Refactor parser to separate files for big functions
