+++ src/parser/src/structs.rs
@@ -1,0 +1,241 @@
+use std::fmt::Display;
+use std::rc::Rc;
+
+use lexer::Token;
+
+#[derive(Debug, Clone)]
+pub enum BinOpType {
+    Add,
+    Sub,
+    Mul,
+    Div,
+}
+
+impl BinOpType {
+    pub fn from_token(token: &Token) -> Result<BinOpType, ParseError> {
+        match token {
+            Token::Plus => Ok(Self::Add),
+            Token::Minus => Ok(Self::Sub),
+            Token::Star => Ok(Self::Mul),
+            Token::Slash => Ok(Self::Div),
+            _ => Err(ParseError::new(&format!(
+                "Expected infix operator but got: {}",
+                token
+            ))),
+        }
+    }
+}
+
+impl Display for BinOpType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let chr = match self {
+            BinOpType::Add => "+",
+            BinOpType::Sub => "-",
+            BinOpType::Mul => "*",
+            BinOpType::Div => "/",
+        };
+        write!(f, "{}", chr)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum UnOpType {
+    Negate,
+    Not,
+}
+
+impl Display for UnOpType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let chr = match self {
+            Self::Negate => "-",
+            Self::Not => "!",
+        };
+
+        write!(f, "{}", chr)
+    }
+}
+
+// Different from bytecode Value because values on op stack might be different (e.g fn call)
+#[derive(Debug, Clone)]
+pub enum Expr {
+    Symbol(String),
+    Integer(i64),
+    Float(f64),
+    Bool(bool),
+    UnOpExpr(UnOpType, Box<Expr>),
+    BinOpExpr(BinOpType, Box<Expr>, Box<Expr>),
+    Block(BlockSeq), // expr can be a block
+}
+
+impl Display for Expr {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let string = match self {
+            Expr::Integer(val) => val.to_string(),
+            Expr::Float(val) => val.to_string(),
+            Expr::Bool(val) => val.to_string(),
+            Expr::UnOpExpr(op, expr) => {
+                format!("({}{})", op, expr)
+            }
+            Expr::BinOpExpr(op, lhs, rhs) => {
+                format!("({}{}{})", lhs, op, rhs)
+            }
+            Expr::Symbol(val) => val.to_string(),
+            Expr::Block(seq) => format!("{{ {} }}", seq),
+        };
+
+        write!(f, "{}", string)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct LetStmtData {
+    pub ident: String,
+    pub expr: Expr,
+    pub type_ann: Option<Type>,
+}
+
+#[derive(Debug, Clone)]
+pub struct AssignStmt {
+    pub ident: String,
+    pub expr: Expr,
+}
+
+impl Display for LetStmtData {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let string = if let Some(ty) = self.type_ann {
+            format!("let {} : {} = {}", self.ident, ty, self.expr)
+        } else {
+            format!("let {} = {}", self.ident, self.expr)
+        };
+
+        write!(f, "{}", string)
+    }
+}
+
+impl Display for AssignStmt {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{} = {}", self.ident, self.expr)
+    }
+}
+
+// Later: LetStmt, IfStmt, FnDef, etc.
+#[derive(Debug, Clone)]
+pub enum Decl {
+    LetStmt(LetStmtData),
+    Assign(AssignStmt),
+    ExprStmt(Expr),
+    Block(BlockSeq),
+}
+
+impl Decl {
+    // Need to clone so we can re-use in pratt parser loop
+    // Reasoning: parsing won't take most of the runtime
+    pub fn to_expr(&self) -> Result<Expr, ParseError> {
+        match self {
+            Self::LetStmt(ref stmt) => {
+                Err(ParseError::new(&format!("'{}' is not an expression", stmt)))
+            }
+            Self::Assign(ref stmt) => {
+                Err(ParseError::new(&format!("'{}' is not an expression", stmt)))
+            }
+            Self::ExprStmt(expr) => Ok(expr.clone()),
+            Self::Block(seq) => Ok(Expr::Block(seq.clone())),
+        }
+    }
+}
+
+impl Display for Decl {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let string = match self {
+            Decl::ExprStmt(expr) => expr.to_string(),
+            Decl::LetStmt(stmt) => stmt.to_string(),
+            Decl::Assign(stmt) => stmt.to_string(),
+            _ => todo!(),
+        };
+
+        write!(f, "{}", string)
+    }
+}
+
+// Last expression is value of program semantics (else Unit type)
+// Program is either one declaration or a sequence of declarations with optional last expression
+#[derive(Debug, Clone)]
+pub struct BlockSeq {
+    pub decls: Vec<Decl>,
+    pub last_expr: Option<Rc<Expr>>,
+}
+
+impl Display for BlockSeq {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let decls = self
+            .decls
+            .iter()
+            .map(|d| d.to_string() + ";")
+            .collect::<String>();
+        let expr = match &self.last_expr {
+            Some(expr) => expr.to_string(),
+            None => String::from(""),
+        };
+
+        write!(f, "{}{}", decls, expr)
+    }
+}
+
+#[derive(Debug, PartialEq)]
+pub struct ParseError {
+    msg: String,
+}
+
+impl ParseError {
+    pub fn new(err: &str) -> ParseError {
+        ParseError {
+            msg: err.to_owned(),
+        }
+    }
+}
+
+impl Display for ParseError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "[ParseError]: {}", self.msg)
+    }
+}
+
+// automatic due to Display
+impl std::error::Error for ParseError {}
+
+// Type annotation corresponding to compile time types
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum Type {
+    Int,
+    Float,
+    Bool,
+    Unit,
+}
+
+impl Type {
+    /// Converts string to primitive type.
+    pub fn from_string(input: &str) -> Result<Type, ParseError> {
+        match input {
+            "int" => Ok(Self::Int),
+            "bool" => Ok(Self::Bool),
+            "float" => Ok(Self::Float),
+            _ => Err(ParseError::new(&format!(
+                "Unknown primitive type: {}",
+                input
+            ))),
+        }
+    }
+}
+
+impl Display for Type {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let string = match self {
+            Self::Int => "int",
+            Self::Bool => "bool",
+            Self::Float => "float",
+            Self::Unit => "()",
+        };
+
+        write!(f, "{}", string)
+    }
+}
